pyproject.toml
[tool.poetry]
name = "qenable"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "qenable", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4"
flask = ">=2.2.2,<2.3.0"
api-client = "^1.3.1"
pyyaml = ">=6.0,<7.0"
pyrdfa3 = "^3.5.3"
python-dotenv = "^0.21.0"
urllib3 = ">=1.26.12,<1.27.0"
qaclient = {path = "qaclient"}
python-scrapyd-api = "^2.1.2"
flask-cors = "^3.0.10"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.6"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
